Add files via upload
The Vehicle Pollution Monitoring System is a DIY project that aims to detect the concentration of harmful gases emitted by vehicles and alert the owner when the pollution levels exceed a threshold limit. The system utilizes an Arduino Uno microcontroller, GSM 800 A module for communication, MQ 107 gas sensor with I2C chip for gas detection, and a 16 x 2 LCD display for real-time information. Once the value is above the desired level. It would send a message as a precautionary measure to the pollution regulatory board.